scripts/requirements.txt
# PMMSherpa Ingestion Pipeline Dependencies
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pyyaml>=6.0
tqdm>=4.65.0
